% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.protoclust.R
\name{as.protoclust}
\alias{as.protoclust}
\title{Convert a generic hclust object to a protoclust object.}
\usage{
as.protoclust(hc, protos)
}
\arguments{
\item{hc}{A generic hclust object using any linkage.}

\item{protos}{A vector of prototype label indices for each branch in the dendrogram.}
}
\description{
Convert a generic hclust object to a protoclust object.
}
