% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{branch}
\alias{branch}
\title{Get an hclust object consisting of the branch with root given by a particular interior node}
\usage{
branch(hc, merge.id)
}
\arguments{
\item{hc}{protoclust object}

\item{merge.id}{a particular interior node}
}
\value{
an hclust object consisting of the branch with root given by merge.id.
}
\description{
Get an hclust object consisting of the branch with root given by a particular interior node
}
\keyword{internal}
