% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{covid19}
\alias{covid19}
\title{Dataset covid19}
\source{
\url{https://cmu-delphi.github.io/covidcast/covidcastR/}
}
\description{
This is a protoclust object for use with protoshiny. It is the hierarchical clustering of US counties using mobility data (proportion of people staying home at least 8 hours and proportion of people going into a restaurant).
}
