% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dynGraph}
\alias{dynGraph}
\title{Dynamic element of class "d3graph"}
\usage{
dynGraph(inputoutputId)
}
\arguments{
\item{inputoutputId}{ID of the input/output element}
}
\description{
Dynamic element of class "d3graph"
}
\keyword{internal}
