% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{faces}
\alias{faces}
\title{Dataset faces}
\source{
\url{https://www.kaggle.com/kasikrit/att-database-of-faces}
}
\description{
This is a protoclust object for use with protoshiny. It is the hierarchical clustering of 400 grayscale images of faces that correspond to 10 images of 40 individuals.
}
