% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fixUploadedFilesNames}
\alias{fixUploadedFilesNames}
\title{Fix uploaded files are getting renamed}
\usage{
fixUploadedFilesNames(x)
}
\arguments{
\item{x}{uploaded object}
}
\description{
Fix uploaded files are getting renamed
}
\keyword{internal}
