% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_paths_from_cut}
\alias{get_paths_from_cut}
\title{Given a clustering get the root-to-node paths for visible nodes}
\usage{
get_paths_from_cut(hc, cl)
}
\arguments{
\item{hc}{An object of class `hclust`}

\item{cl}{A numeric vector in the format of the output of `stats::cutree` 
that gives the cluster memberships of each leaf.}
}
\description{
This function returns the binary path from root to node for each
node that should be displayed in order to return the input clustering.
Note that this  will only be possible for a clustering `cl` in which each 
cluster corresponds to a branch of `hc`.
}
