% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_obj}
\alias{load_obj}
\title{Load an object into a new environment}
\usage{
load_obj(file)
}
\arguments{
\item{file}{object to load into the new environment}
}
\description{
Load an object into a new environment
}
\keyword{internal}
