% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{movies}
\alias{movies}
\title{Dataset movies}
\source{
\url{https://grouplens.org/datasets/movielens/25m/}
}
\description{
This is a protoclust object for use with protoshiny. It is the hierarchical clustering of 13,816 movies based on users' ratings and taggings of movies.
}
