% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{protoclust_to_json}
\alias{protoclust_to_json}
\title{Convert a protoclust object to a JSON tree}
\usage{
protoclust_to_json(proto_object)
}
\arguments{
\item{proto_object}{protoclust object including merge, labels, and height}
}
\description{
Convert a protoclust object to a JSON tree
}
\keyword{internal}
