% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{seaflow}
\alias{seaflow}
\title{Dataset seaflow}
\source{
Provided by Mattias Rolf Cape and Francois Ribalet.
}
\description{
This is a protoclust object for use with protoshiny. It is the hierarchical clustering of 6,336 scatterplots resulting from continuous-time flow cytometry measurements of phytoplankton.
}
