% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{selectJS}
\alias{selectJS}
\title{Dynamic element of class "select_custom"}
\usage{
selectJS(inputoutputId)
}
\arguments{
\item{inputoutputId}{ID of the input/output element}
}
\description{
Dynamic element of class "select_custom"
}
\keyword{internal}
