% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{visualize_hc}
\alias{visualize_hc}
\title{Run a local instance of protoshiny}
\usage{
visualize_hc()
}
\value{
No return value, launches the protoshiny Shiny application.
}
\description{
Run a local instance of protoshiny
}
\examples{
\dontrun{
    visualize_hc()
}
}
