% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{words}
\alias{words}
\title{Dataset words}
\source{
\url{https://cs.nyu.edu/~roweis/data.html}
}
\description{
This is a protoclust object for use with protoshiny. It is the hierarchical clustering of 19,599 of the most common words from Grolier's Encyclopedia.
}
