% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Extract cluster labels}
\usage{
get_clusters(hc)
}
\arguments{
\item{hc}{A protoshiny object downloaded from the web interface.}
}
\value{
A vector of saved cluster assignments resulting from interaction in the protoshiny application.
}
\description{
Extract cluster labels
}
